+++ modules/profiler/index.js
@@ -3,19 +3,17 @@
 var gracenode = require('gracenode');
 var log = gracenode.log.create('profiler');
 
-module.exports.create = function (name, logType) {
-	return new Profiler(name, logType);
+module.exports.create = function (name) {
+	return new Profiler(name);
 };
 
-// logType: verbose, debug, info, warning, error, fatal > default is verbose
-function Profiler(name, logType) {
+function Profiler(name) {
 	this._name = name;
 	this._startTime = 0;
 	this._nowTime = 0;
 	this._marks = [];
 	this._running = false;
-	this._type = logType || 'verbose';
-	this._enabled = gracenode.log.isEnabled(this._type);
+	this._enabled = gracenode.log.isEnabled('debug');
 	log.verbose('profiler [' + name + '] with', this._type, 'enabled:', this._enabled);
 }
 
@@ -29,7 +27,7 @@ Profiler.prototype.start = function () {
 		return log.warning('profiler is currently running. invoke "stop" before calling "start"');
 	}
 	
-	log[this._type]('profiling of "' + this._name + '" started');
+	log.debug('profiling of "' + this._name + '" started');
 
 	var date = new Date();
 	this._startTime = date.getTime();
@@ -67,57 +65,28 @@ Profiler.prototype.stop = function () {
 		return log.warning('stop called, but profiler "' + this._name + '" is not running');
 	}
 
-	log[this._type]('profiling of "' + this._name + '" stopped');
-
-	var date = new Date();
-	var now = date.getTime();
-	var totalTime = now - this._startTime;
-	var logList = [];
-	var len = this._marks.length;
-	var gap = 0;
-	var space = '';
-	var msg = ' ' + this._name + ' took [' + totalTime + ' ms] ';
-	var msgLen = msg.length;
-	var longest = msgLen;
-	// degenrate profiling logs
-	for (var i = 0; i < len; i++) {
-		if (!this._marks[i].name) {
-			// if there is no message, skip
-			continue;
-		}
+	log.debug('profiling of "' + this._name + '" stopped');
+	
+	var totalTime = Date.now() - this._startTime;
+	var table = [];
+	for (var i = 0, len = this._marks.length; i < len; i++) {
 		var item = this._marks[i];
-		var str = ' > ' + item.name + ' took [' + item.time + ' ms] ';
-		if (longest < str.length) {
-			longest = str.length;
-		}
-		logList.push(str);
-	}
-	var line = '+';
-	for (var a = 0; a < longest; a++) {
-		line += '-';
-	}
-	line += '+';
-	log[this._type](line);
-	for (var j = 0; j < len; j++) {
-		if (!logList[j]) {
-			// this is a skipped log
+		if (!item.name) {
+			// no name provided, skip it
 			continue;
 		}
-		gap = longest - logList[j].length;
-		space = '';
-		for (var n = 0; n < gap; n++) {
-			space += ' ';
-		}
-		log[this._type]('|' + logList[j] + space + '|');
-		log[this._type](line);
+		table.push({
+			name: item.name,
+			'execution time': item.time + ' milliseconds'
+		});
 	}
-	gap = longest - msgLen;
-	space = '';
-	for (var b = 0; b < gap; b++) {
-		space += ' ';
-	}
-	log[this._type]('|' + msg + space + '|');
-	log[this._type](line);
+	// add total execution time
+	table.push({
+		name: 'total execution time',
+		'execution time': totalTime + ' milliseconds'
+	});
+	log.table(table);
+
 	// flush out and reset
 	this._startTime = 0;
 	this._nowTime = 0;