+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This is a list of manually mantained changes and updates for each version.
 
 #### Updated a dependency version
 
-Updated gracelog version from `0.1.13` to `0.1.17`.
+Updated gracelog version from `0.1.13` to `0.1.18`.
 
 ## Deprecated